p/work log & work plan completed
